module12/requirements.txt
google-cloud-spanner
networkx